BACKEND-FASTAPI/requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching & Queue
redis==5.0.1
celery==5.3.6
flower==2.0.1

# Security & Crypto
cryptography==42.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pycryptodome==3.20.0

# Post-Quantum Crypto (CRITICAL)
oqs-python==0.9.0
pqcrypto==0.1.7

# Data Validation
pydantic==2.6.0
pydantic-settings==2.2.0

# Real-time
websockets==12.0
python-socketio==5.11.1

# Notifications
sendgrid==6.11.0
twilio==9.0.4
python-telegram-bot==20.8

# File Processing
openpyxl==3.1.2
reportlab==4.1.0
pillow==10.2.0

# Utilities
langdetect==1.0.9
python-dateutil==2.9.0
pytz==2024.1

# Monitoring
prometheus-client==0.20.0
sentry-sdk==1.40.6

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
faker==24.0.0

# Development
black==24.2.0
flake8==7.0.0
mypy==1.8.0
